+++ src/config/imap.php
@@ -50,7 +50,7 @@ return [
             'host'  => env('IMAP_HOST', 'localhost'),
             'port'  => env('IMAP_PORT', 993),
             'protocol'  => env('IMAP_PROTOCOL', 'imap'), //might also use imap, [pop3 or nntp (untested)]
-            'encryption'    => env('IMAP_ENCRYPTION', 'ssl'), // Supported: false, 'ssl', 'tls', 'notls'
+            'encryption'    => env('IMAP_ENCRYPTION', 'ssl'), // Supported: false, 'ssl', 'tls', 'notls', 'starttls'
             'validate_cert' => env('IMAP_VALIDATE_CERT', true),
             'username' => env('IMAP_USERNAME', 'root@example.com'),
             'password' => env('IMAP_PASSWORD', ''),
@@ -60,7 +60,7 @@ return [
         'gmail' => [ // account identifier
             'host' => 'imap.gmail.com',
             'port' => 993,
-            'encryption' => 'ssl', // Supported: false, 'ssl', 'tls'
+            'encryption' => 'ssl',
             'validate_cert' => true,
             'username' => 'example@gmail.com',
             'password' => 'PASSWORD',
@@ -69,7 +69,7 @@ return [
         'another' => [ // account identifier
             'host' => '',
             'port' => 993,
-            'encryption' => false, // Supported: false, 'ssl', 'tls'
+            'encryption' => false,
             'validate_cert' => true,
             'username' => '',
             'password' => '',